Moved the exisitng index to /canvas (to spare index for landing page), edit meta data so they have the same format, hide header and sidebar from index
