+++ pages/index.tsx
@@ -1,55 +1,10 @@
-import { collection, getDocs } from '@firebase/firestore';
-import { InferGetStaticPropsType } from 'next';
-import dynamic from 'next/dynamic';
+import { Container, Typography } from '@mui/material';
+import { Inter } from '@next/font/google';
 import Head from 'next/head';
-import { useEffect } from 'react';
-import { Background, Frame, Poster } from '../components/types';
-import { useSidebar } from '../context/SidebarContext';
-import { db } from '../firebase/firebaseConfig';
 
-const Test = dynamic(() => import('../components/canvas/Test'), { ssr: false }); // do not adjust this - M1 mac needs this to run canvas
+const inter = Inter({ subsets: ['latin'] });
 
-export const getStaticProps = async () => {
-  const framesCollectionRef = collection(db, 'frames');
-  const frameData = await getDocs(framesCollectionRef);
-  const backgroundsCollectionRef = collection(db, 'backgrounds');
-  const backgroundData = await getDocs(backgroundsCollectionRef);
-  const postersCollectionRef = collection(db, 'posters');
-  const posterData = await getDocs(postersCollectionRef);
-
-  return {
-    props: {
-      frames: frameData.docs.map((doc) => ({
-        ...(doc.data() as Frame),
-        id: doc.id,
-      })),
-      backgrounds: backgroundData.docs.map((doc) => ({
-        ...(doc.data() as Background),
-        id: doc.id,
-        createdAt: doc.data().createdAt.toDate().toDateString(),
-      })),
-      posters: posterData.docs.map((doc) => ({
-        ...(doc.data() as Poster),
-        id: doc.id,
-        createdAt: doc.data().createdAt.toDate().toDateString(),
-      })),
-    },
-  };
-};
-
-export default function Home({
-  frames,
-  backgrounds,
-  posters,
-}: InferGetStaticPropsType<typeof getStaticProps>) {
-  const { setAllFrames, setAllBackgrounds, setAllPosters } = useSidebar();
-  useEffect(
-    () => setAllBackgrounds(backgrounds),
-    [backgrounds, setAllBackgrounds]
-  );
-  useEffect(() => setAllFrames(frames), [frames, setAllFrames]);
-  useEffect(() => setAllPosters(posters), [posters, setAllPosters]);
-
+export default function Home() {
   return (
     <>
       <Head>
@@ -58,7 +13,11 @@ export default function Home({
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <Test />
+      <main>
+        <Container>
+          <Typography>Welcome to Blue print!</Typography>
+        </Container>
+      </main>
     </>
   );
 }
